+++ Sources/StreamChat/Models/Device+Extensions.swift
@@ -1,0 +1,15 @@
+//
+// Copyright © 2026 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+
+/// A unique identifier of a device.
+public typealias DeviceId = String
+
+extension Data {
+    /// Generates a device id string from device token data.
+    var deviceId: DeviceId { map { String(format: "%02x", $0) }.joined() }
+}
+
+extension Device: Identifiable {}